Added video playback. Made building with Kinect SDK optional. Fixed windows build, warnings.

+++ Software/Tracker/src/comm/datahandler.cpp
@@ -23,7 +23,7 @@ void DataHandler_TASK(uint16_t elapsedMs)
 			if (transmitPacket.ulPeriod > 0 && packetDesc->iPeriod >= 0)
 			{
 				packetDesc->iPeriod += elapsedMs;
-				if (packetDesc->iPeriod > transmitPacket.ulPeriod)
+				if (uint32_t(packetDesc->iPeriod) > transmitPacket.ulPeriod)
 					packetDesc->iPeriod = elapsedMs;
 			}
 		}
